+++ training/include/card_buckets.h
@@ -13,11 +13,11 @@ static size_t count_clusters(const std::vector<T>& assign) {
 }
 
 struct CardBuckets {
-    std::vector<size_t> preflop_clusters;
-    std::vector<size_t> flop_clusters;
-    std::vector<size_t> turn_clusters;
-    std::vector<size_t> river_clusters;
-    std::vector<size_t> cluster_counts;
+    std::vector<int> preflop_clusters;
+    std::vector<int> flop_clusters;
+    std::vector<int> turn_clusters;
+    std::vector<int> river_clusters;
+    std::vector<int> cluster_counts;
 
     CardBuckets() = default; 
 
@@ -29,20 +29,20 @@ struct CardBuckets {
 
     void set_clusters(const std::string& flop_path, const std::string&turn_path, const std::string& river_path){
 
-        size_t num_preflops = 169;
+        int num_preflops = 169;
         preflop_clusters.resize(num_preflops);
     
-        for (size_t i = 0; i < num_preflops; ++i){
+        for (int i = 0; i < num_preflops; ++i){
             preflop_clusters[i] = static_cast<int>(i);
         }
 
-        auto [fc, flop_header] = load_matrix_and_header<size_t>(flop_path);
+        auto [fc, flop_header] = load_matrix_and_header<int>(flop_path);
         flop_clusters = std::move(fc);
 
-        auto[tc, turn_header] = load_matrix_and_header<size_t>(turn_path);
+        auto[tc, turn_header] = load_matrix_and_header<int>(turn_path);
         turn_clusters = std::move(tc);
 
-        auto [rc, river_header] = load_matrix_and_header<size_t>(river_path);
+        auto [rc, river_header] = load_matrix_and_header<int>(river_path);
         river_clusters = std::move(rc);
 
         cluster_counts.clear();
@@ -52,7 +52,7 @@ struct CardBuckets {
         cluster_counts.push_back(count_clusters(river_clusters));
     }
 
-    size_t cluster_of(int street, int hand_id) const {
+    int cluster_of(int street, int hand_id) const {
         if (street == 0) return preflop_clusters[hand_id];
         if (street == 1) return flop_clusters[hand_id];
         if (street == 2) return turn_clusters[hand_id];
